# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API сервиса многостороннего обмена вещами: профили пользователей, вход по JWT\nи объявления о вещах, которые владелец готов обменять.\n\nЗащищённые маршруты читают HttpOnly cookie `access_token` — они помечены замком.\nКнопки «Authorize» здесь нет и не нужно: выполните `POST /auth/login` прямо из этой\nстраницы — браузер сохранит cookie и будет отправлять её со всеми следующими\nзапросами сам. JavaScript до неё не дотянется, поэтому вписать её руками нельзя.",
        "title": "Цепочка обмена — API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/exchanges/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Закрывает proposed или confirmed обмен. Для confirmed освобождает объявления и запускает повторный поиск. Решения участников и их статистика не изменяются.",
                "tags": [
                    "admin"
                ],
                "summary": "Принудительно отменить обмен",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID обмена",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Обмен отменён"
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Обмен не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Обмен уже завершён или отменён либо объявления находятся в несовместимом состоянии",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/exchanges": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает все найденные обмены текущего пользователя вместе с участниками и объявлениями.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Получить свои обмены",
                "responses": {
                    "200": {
                        "description": "Список обменов; если обменов нет, возвращается []",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ExchangeResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exchanges/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Доступен только участнику этого обмена.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Получить обмен по ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "8db9f3e2-8a45-4a70-b3d1-167b4f97e121",
                        "description": "UUID обмена",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обмен",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Пользователь не участвует в обмене",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Обмен не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/complete": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Сохраняет подтверждение текущего участника. После подтверждения всеми участниками обмен завершается, объявления помечаются переданными, а счётчики завершённых обменов увеличиваются.",
                "tags": [
                    "exchanges"
                ],
                "summary": "Подтвердить получение вещи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID обмена",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Получение подтверждено"
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Пользователь не участвует в обмене",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Обмен не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Обмен ещё не подтверждён или уже отменён",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exchanges/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Сохраняет согласие текущего участника. После согласия всех участников обмен подтверждается, а объявления резервируются.",
                "tags": [
                    "exchanges"
                ],
                "summary": "Подтвердить участие в обмене",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID обмена",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Участие подтверждено"
                    },
                    "400": {
                        "description": "ID не является UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Пользователь не участвует в обмене",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Обмен не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Решение уже принято, обмен закрыт или объявление недоступно",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ParticipantItemResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.CategoryResponse"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pickup_point": {
                    "description": "null — вещь ещё дома у владельца. По этому полю видно, кого из участников ждут.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ParticipantPickupPointResponse"
                        }
                    ],
                    "x-nullable": true
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.ParticipantResponse": {
            "type": "object",
            "properties": {
                "completion_confirmed_at": {
                    "type": "string"
                },
                "decided_at": {
                    "type": "string"
                },
                "gives_item": {
                    "$ref": "#/definitions/dto.ParticipantItemResponse"
                },
                "position": {
                    "type": "integer"
                },
                "receives_item": {
                    "$ref": "#/definitions/dto.ParticipantItemResponse"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.ParticipantUserResponse"
                }
            }
        },
        "dto.ParticipantUserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_sweetlife999_chain-of-trades-avito_internal_exchange_dto.ExchangeResponse": {
            "type": "object",
            "properties": {
                "cancel_reason": {
                    "type": "string",
                    "enum": [
                        "proposal_declined",
                        "confirmed_broken",
                        "superseded",
                        "item_withdrawn",
                        "user_blocked",
                        "admin_cancelled",
                        "legacy"
                    ],
                    "x-nullable": true
                },
                "closed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ParticipantResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "unread_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "HttpOnly cookie, которую ставит POST /auth/login.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}